# Conflicts:
#	api/services.py

+++ api/services.py
@@ -1,3 +1,4 @@
+import re
 from datetime import datetime
 from api.models import Course, Classroom, Instructor
 import logging
@@ -42,10 +43,12 @@ def calculate_time_blocks(buildings):
             # Finds all start/end times in the specified building on the current day
             start_times = [time[0].strftime("%H:%M:%S") for time in
                            Course.objects.all().filter(classroom__building__in=buildings, day__contains=day).
-                           values_list('start_time').distinct().exclude(start_time__isnull=True)]
+                           values_list('start_time').distinct().exclude(start_time__isnull=True)
+                           .exclude(classroom__building__in=["Unknown", "OFCP"])]
             end_times = [time[0].strftime("%H:%M:%S") for time in
                          Course.objects.all().filter(classroom__building__in=buildings, day__contains=day).
-                         values_list('end_time').distinct().exclude(end_time__isnull=True)]
+                         values_list('end_time').distinct().exclude(end_time__isnull=True)
+                         .exclude(classroom__building__in=["Unknown", "OFCP"])]
 
         all_times = start_times + end_times + ['06:00:00', '23:59:00']
         start_end_times = sorted(set(all_times))  # Organizes times from earliest to latest
@@ -124,7 +127,7 @@ def get_all_buildings():
     return buildings_list
 
 
-def get_used_classrooms(day: str, start_time: str, end_time: str, buildings: str = "all") -> {}:
+def get_used_classrooms(day: str, start_time: str, end_time: str, buildings: [] = "all") -> {}:
     """
     Returns a list of all classrooms used during a specified time block and data about the course being held in the
     classroom during that specified time block.
@@ -133,8 +136,14 @@ def get_used_classrooms(day: str, start_time: str, end_time: str, buildings: str
     :param start_time: string specifying the start time in which to search for used classrooms
     :param end_time:   string specifying the time in which searching for used classrooms stops
     :param buildings:  string indicating which buildings should be searched for used classrooms
-    :return            dictionary containing the data for every classroom being used within the specified time block and building
+    :return            dictionary containing the data for every classroom being used within the specified time block and
+                       building
     """
+    # Start/end times must be in HH:MM:SS format
+    if not re.match(r'^\d{2}:\d{2}$', start_time) or not re.match(r'^\d{2}:\d{2}$', end_time):
+        logger.debug("Returning empty dictionary due to invalid start/end times")
+        return {}
+
     if buildings == 'all':
         # Searches for used classrooms within all buildings
         current_courses = (Course.objects.all().filter(day__contains=day,
@@ -166,7 +175,7 @@ def get_used_classrooms(day: str, start_time: str, end_time: str, buildings: str
     return classrooms_dict
 
 
-def calculate_classroom_time_blocks(classroom):
+def calculate_classroom_time_blocks(classroom: str):
     """
     Finds all possible time blocks used in the specified classroom and then returns this information in a dictionary.
     This dictionary uses the abbreviation for every weekday ('M', 'T', etc.) as the keys and an array of arrays as
@@ -246,10 +255,25 @@ def get_past_time(day, current_time, buildings='all'):
     :param current_time: string specifying the end time for a block; the paired start time is the desired output
     :return              string detailing the start time associated with the specified end time
     """
+    # The day must be M,T,W,th, or F
+    if day not in ['M', 'T', 'W', 'th', 'F']:
+        logger.debug(f"{day} is not a valid day")
+        return ''
+    # Convert HH:MM format to HH:MM:SS format
+    if re.match(r'^\d{2}:\d{2}$', current_time):
+        logger.debug("Time format changed to HH:MM:SS")
+        current_time = f"{current_time}:00"
+    # The time must be in HH:MM:SS format for comparison with the times from calculate_time_blocks()
+    if not re.match(r'^\d{2}:\d{2}:\d{2}$', current_time):
+        logger.debug(f"{current_time} is not a valid time")
+        return ''
     time_blocks = calculate_time_blocks(buildings)
     for block in time_blocks[day]:
-        if datetime.strptime(block[1], "%H:%M:%S").time() == datetime.strptime(current_time, "%H:%M").time():
+        if block[1] == current_time:
             return block[0][:-3]
+    # If the time is not found within the time blocks list, return empty string
+    logger.debug(f"{current_time} not found within the specified time blocks")
+    return ''
 
 
 def get_next_time(day, current_time, buildings='all'):
@@ -262,12 +286,27 @@ def get_next_time(day, current_time, buildings='all'):
     :param current_time: string specifying the start time which starts the block; the paired end time is the desired output
     :return              string detailing the end time associated with the specified start time
     """
+    # The day must be M,T,W,th, or F
+    if day not in ['M', 'T', 'W', 'th', 'F']:
+        logger.debug(f"{day} is not a valid day")
+        return ''
+    # Convert HH:MM format to HH:MM:SS format
+    if re.match(r'^\d{2}:\d{2}$', current_time):
+        logger.debug("Time format changed to HH:MM:SS")
+        current_time = f"{current_time}:00"
+    # The time must be in HH:MM:SS format for comparison with the times from calculate_time_blocks()
+    if not re.match(r'^\d{2}:\d{2}:\d{2}$', current_time):
+        logger.debug(f"{current_time} is not a valid time")
+        return ''
     time_blocks = calculate_time_blocks(buildings)
     logger.debug(f"get_next_time: Time Blocks: {time_blocks}")
     for block in time_blocks[day]:
-        if datetime.strptime(block[0], "%H:%M:%S").time() == datetime.strptime(current_time, "%H:%M").time():
+        if block[0] == current_time:
             logger.debug(f"get_next_time: Next Time Found: {block[1][:-3]}")
             return block[1][:-3]
+    # If the time is not found within the time blocks list, return empty string
+    logger.debug(f"{current_time} not found within the specified time blocks")
+    return ''
 
 
 def calculate_day_string(row):
@@ -299,6 +338,10 @@ def upload_schedule_data(file):
 
     :param file: Excel spreadsheet containing the scheduled course data for populating the database
     """
+    # The file must be an Excel spreadsheet to be uploaded to the DB
+    if file[-5:] != '.xlsx':
+        return
+
     # Delete all data to prevent different semesters being present in the DB
     Course.objects.all().delete()
     Classroom.objects.all().delete()
@@ -306,11 +349,6 @@ def upload_schedule_data(file):
 
     logger.info("New schedule file uploading")
     df = pd.read_excel(file)
-
-    # Delete all courses already in the database for the term being modified
-    current_term = df['SEC_TERM'][0]
-    current_term_courses = Course.objects.all().filter(term__exact=current_term)
-    current_term_courses.delete()
 
     # Create new courses from the uploaded file
     for index, row in df.iterrows():
@@ -344,8 +382,8 @@ def upload_schedule_data(file):
             status=None if pd.isna(row['SEC_STATUS']) else row['SEC_STATUS'],
             start_time=None if pd.isna(row['CSM_START_TIME']) else datetime.strptime(row['CSM_START_TIME'],
                                                                                      '%I:%M%p').time(),
-            end_time=None if pd.isna(row['CSM_START_TIME']) else datetime.strptime(row['CSM_END_TIME'],
-                                                                                   '%I:%M%p').time(),
+            end_time=None if pd.isna(row['CSM_END_TIME']) else datetime.strptime(row['CSM_END_TIME'],
+                                                                                 '%I:%M%p').time(),
             day=day_string,
             classroom=classroom if classroom is not None else None,
             instruction_method=row['CSM_INSTR_METHOD'],
@@ -363,6 +401,11 @@ def upload_classroom_data(file):
 
     :param file: Excel spreadsheet containing the classroom specification data for populating the database
     """
+    # The file must be an Excel spreadsheet
+    if file[-5:] != '.xlsx':
+        logger.debug("Wrong file type entered")
+        return
+
     logger.info("New classroom data uploading")
     df = pd.read_excel(file)
 
@@ -386,7 +429,7 @@ def upload_classroom_data(file):
         logger.debug(f"upload_classroom_data: Classroom Features: {features}")
 
         # Create the new updated classroom object
-        classroom = Classroom.objects.update_or_create(
+        classroom, _ = Classroom.objects.update_or_create(
             name=correct_bldg_names_dict[row['Building Information']] + "-" + str(row['Room Number']),
             building=correct_bldg_names_dict[row['Building Information']],
             room_num=row['Room Number'],
